requirements.txt
django
django-extensions
nltk
psycopg2-binary
pandas
numpy
pysrt
django-debug-toolbar
requests
django-cors-headers
redis
django-redis-cache
django-redis
djangorestframework
pillow
django-rq
gunicorn
black
gevent
huey
django-admin-autocomplete-filter